chore: cleanup tex document related providers (pt.1)
